globally renamed uint to size_t. This has _not_ been checked under Windows, although I changed msw files also, so please wait until this evening if you want to be sure that it compiles. This change should fix 64 bit compilation problems, but it would be nice to test it...
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@591 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/config.h
@@ -123,8 +123,8 @@ public:
   virtual bool GetNextEntry (wxString& str, long& lIndex) const = 0;
     // get number of entries/subgroups in the current group, with or without
     // it's subgroups
-  virtual uint GetNumberOfEntries(bool bRecursive = FALSE) const = 0;
-  virtual uint GetNumberOfGroups(bool bRecursive = FALSE) const = 0;
+  virtual size_t GetNumberOfEntries(bool bRecursive = FALSE) const = 0;
+  virtual size_t GetNumberOfGroups(bool bRecursive = FALSE) const = 0;
 
   // tests of existence
     // returns TRUE if the group by this name exists
